Comments about political opinion are offensive
Legislation in Northern Ireland treats discrimination on grounds of political opinion similarly to the grounds of religion. This change amends the code of conduct to include political opinion and to prioritise neither of these two grounds of discrimination. For a definition of political opinion see for example:  

http://www.equalityni.org/Individuals/I-have-a-problem-with-a-service/Religious-belief-Political-opinion

> ‘Political opinion’ includes support of any Northern Ireland or UK government politics and matters of public policy.  For example:  Treated less favourably because you are or perceived to be nationalist or unionist. However, not all political opinions have the protection of the law. It does not protect political opinions that support or approve of the use of violence for political ends.

or http://www.legislation.gov.uk/nisi/1998/3162/made