Insert an accountability section
Put together a plan for enforcement-accountability, as this can be much harder with a non-discursive, constitution-based reference point. This would help limit concern that unconscious bias will affect how the CoC is applied.

We haven't had a "you may disagree but it says so in the CoC" or equivalent as effective discussion-enders before, but that's only valid when there's some accountability around when and how consistently that line is (implicitly/explicitly) used, and who is affected positively/negatively over time.