Update Spam definition to include non-individual accounts
While there has never been a 'real-name' policy, there has been a historical consensus that brand-accounts are not appropriate for this community space. 

Personal handles/monitors/aliases etc continue to be acceptable; however, accounts that purely represent corporate, organisation or brand identity are not appropriate and should be considered spam.